+++ src/app/components/controls/AddButton.jsx
@@ -2,13 +2,7 @@ import React from "react";
 
 export const AddButton = (props) => {
 
-    let {onClick} = props
-
-    const handleClick = (e) => {
-        onClick && onClick(e)
-    }
-
     return (
-        <button className="Add-button" onClick={handleClick}>+</button>
+        <button className="Add-button" {...props}>+</button>
     )
 }