+++ src/app/slices/api/eventApiSlice.js
@@ -1,6 +1,6 @@
 import {apiSlice, competitionTag, eventTag} from "./apiSlice";
 import store from "../../store";
-import {matchAdapter, matchLoaded, matchSlice} from "../matchSlice";
+import {artworkRefreshed, matchAdapter, matchDeleted, matchLoaded, matchSlice} from "../matchSlice";
 import {JsonHeaders} from "../../constants";
 
 
@@ -86,6 +86,14 @@ export const eventApiSlice = apiSlice.injectEndpoints({
                     method: 'POST',
                 }),
                 invalidatesTags: [eventTag],
+                transformResponse: response => {
+                    let {_links: links} = response
+                    if (links) {
+                        let {image} = links
+                        store.dispatch(artworkRefreshed(image?.href))
+                    }
+                    return response
+                },
             }),
             addMatch: builder.mutation({
                 query: match => ({
@@ -111,6 +119,10 @@ export const eventApiSlice = apiSlice.injectEndpoints({
                     method: 'DELETE',
                 }),
                 invalidatesTags: [eventTag, competitionTag],
+                transformResponse: matchId => {
+                    store.dispatch(matchDeleted(matchId))
+                    return matchId
+                },
             }),
         })
     }